doubleszz

+++ algoriddim.h
@@ -8,17 +8,17 @@ typedef struct t_algoriddim
 } algo;
 
 typedef struct melody_msg {
-  int *melody;
+  double *melody;
   int osc_num;
   int melody_note_len;
   int melody_loop_len;
   int melody_play_lock;
   int melody_cur_note;
-  int mod_freq; // modulator for an FM message
+  double mod_freq; // modulator for an FM message
 } melody_msg;
 
 
-melody_msg* new_melody_msg(int *freqs, int melody_note_len, int loop_len);
+melody_msg* new_melody_msg(double *freqs, int melody_note_len, int loop_len);
 
 algo *new_algo(void);
 
@@ -26,6 +26,6 @@ void *algo_run(void *);
 void *loop_run(void *);
 void *floop_run(void *);
 void *randdrum_run(void *m);
-void play_melody(const int osc_num, int *mlock, int *note, int *notes, const int note_len);
-void fplay_melody(const int sg_num, int *mlock, int *note, int *notes, const int note_len);
+void play_melody(const int osc_num, int *mlock, int *note, double *notes, const int note_len);
+void fplay_melody(const int sg_num, int *mlock, int *note, double *notes, const int note_len);
 void mk_sbmsg_sine(int freq);
