+++ mixer.h
@@ -41,8 +41,8 @@ typedef struct {
 mixer *new_mixer(void);
 
 void mixer_ps(mixer *mixr);
-int add_osc(mixer *mixr, int freq, GTABLE *gt);
-int add_fm(mixer *mixr, int ffreq, int cfreq);
+int add_osc(mixer *mixr, double freq, GTABLE *gt);
+int add_fm(mixer *mixr, double ffreq, double cfreq);
 int add_drum(mixer *mixr, char *filename, char *pattern);
 int add_sampler(mixer *mixr, char *filename, int loop_len);
 int add_sound_generator(mixer *mixr, SBMSG *sbm);
